requirements.txt
aws-cdk-lib==2.162.1
constructs>=10.0.0,<11.0.0
boto3==1.26.137
requests==2.28.2
python-dotenv
openai==1.12.0
httpx==0.27.0
pydantic==2.6.1
pydantic-core==2.16.2
typing-extensions==4.9.0
